+++ lib/rspec/mocks/argument_list_matcher.rb
@@ -44,6 +44,7 @@ module RSpec
       # @see #args_match?
       def initialize(*expected_args)
         @expected_args = expected_args
+        ensure_expected_args_valid!
       end
 
       # @api public
@@ -56,11 +57,6 @@ module RSpec
       def args_match?(*args)
         Support::FuzzyMatcher.values_match?(matchers_for(args), args)
       end
-
-      # Value that will match all argument lists.
-      #
-      # @private
-      MATCH_ALL = new(ArgumentMatchers::AnyArgsMatcher::INSTANCE)
 
     private
 
@@ -80,6 +76,22 @@ module RSpec
         any_args = 1.upto(any_args_count).map { ArgumentMatchers::AnyArgMatcher::INSTANCE }
         expected_args.first(before_count) + any_args + expected_args.last(after_count)
       end
+
+      def ensure_expected_args_valid!
+        if expected_args.count(ArgumentMatchers::AnyArgsMatcher::INSTANCE) > 1
+          raise ArgumentError, "`any_args` can only be passed to " \
+                "`with` once but you have passed it multiple times."
+        elsif expected_args.count > 1 && expected_args.include?(ArgumentMatchers::NoArgsMatcher::INSTANCE)
+          raise ArgumentError, "`no_args` can only be passed as a " \
+                "singleton argument to `with` (i.e. `with(no_args)`), " \
+                "but you have passed additional arguments."
+        end
+      end
+
+      # Value that will match all argument lists.
+      #
+      # @private
+      MATCH_ALL = new(ArgumentMatchers::AnyArgsMatcher::INSTANCE)
     end
   end
 end